requirements.txt
Django>=5.0,<6.0
gunicorn>=22.0
whitenoise>=6.7
python-dotenv>=1.0
